requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
Deprecated==1.2.9
discord==1.0.1
discord.py==1.3.3
idna==2.9
multidict==4.7.5
pyaml==20.4.0
PyGithub==1.47
PyJWT==1.7.1
PyYAML==5.3.1
requests==2.23.0
urllib3==1.25.9
websockets==8.1
wrapt==1.12.1
yarl==1.4.2
